+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LASRV010 - 03/04/2023 - Add the RequestManager class.
 - LASRV010 - 03/04/2023 - Add the RequestCPU class.
 - LASRV010 - 03/04/2023 - Add the RequestRAM class.
+- LASRV010 - 04/04/2023 - Add the RequestGPU class.
 
 ### Changed
 - LASRV010 - 03/04/2023 - Changes in Request class.